@@ -20,50 +20,50 @@
     "flow": "$(npm bin)/flow"
   },
   "dependencies": {
-    "@commitlint/cli": "^8.3.4",
-    "@commitlint/config-conventional": "^8.3.4",
+    "@commitlint/cli": "^9.1.2",
+    "@commitlint/config-conventional": "^9.1.2",
     "@rowno/sparkline": "^4.0.0",
-    "antd": "^3.26.6",
-    "apollo-boost": "^0.4.7",
-    "apollo-cache-inmemory": "^1.6.5",
-    "apollo-client": "^2.6.8",
-    "apollo-link": "^1.2.13",
-    "apollo-link-context": "^1.0.19",
-    "apollo-link-error": "^1.1.12",
-    "apollo-link-http": "^1.5.16",
-    "auth0-js": "^9.12.1",
-    "bootstrap-css-only": "^4.3.1",
+    "antd": "^3.26.18",
+    "apollo-boost": "^0.4.9",
+    "apollo-cache-inmemory": "^1.6.6",
+    "apollo-client": "^2.6.10",
+    "apollo-link": "^1.2.14",
+    "apollo-link-context": "^1.0.20",
+    "apollo-link-error": "^1.1.13",
+    "apollo-link-http": "^1.5.17",
+    "auth0-js": "^9.14.0",
+    "bootstrap-css-only": "^4.4.1",
     "classnames": "^2.2.6",
-    "commitizen": "^4.0.3",
-    "commitlint": "^8.3.4",
-    "connected-react-router": "^6.6.1",
-    "cz-conventional-changelog": "^3.0.2",
+    "commitizen": "^4.2.1",
+    "commitlint": "^9.1.2",
+    "connected-react-router": "^6.8.0",
+    "cz-conventional-changelog": "^3.3.0",
     "enquire-js": "^0.2.1",
-    "eslint": "^6.8.0",
-    "flow-bin": "^0.104.0",
-    "graphql": "^14.5.8",
-    "graphql-tag": "^2.10.1",
-    "history": "^4.10.1",
+    "eslint": "^7.8.1",
+    "flow-bin": "^0.133.0",
+    "graphql": "^15.3.0",
+    "graphql-tag": "^2.11.0",
+    "history": "^5.0.0",
     "jwt-decode": "^2.2.0",
-    "lodash": "^4.17.11",
-    "luxon": "^1.21.3",
+    "lodash": "^4.17.20",
+    "luxon": "^1.25.0",
     "nprogress": "^0.2.0",
     "peity-react": "^0.0.7",
-    "qs": "^6.9.1",
-    "rc-drawer": "^2.1.1",
-    "react": "^16.12.0",
-    "react-apollo": "^2.5.8",
+    "qs": "^6.9.4",
+    "rc-drawer": "^4.1.0",
+    "react": "^16.13.1",
+    "react-apollo": "^3.1.5",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^16.12.0",
+    "react-dom": "^16.13.1",
     "react-editmask": "^4.0.3",
-    "react-helmet": "^5.2.1",
-    "react-intl": "^2.9.0",
+    "react-helmet": "^6.1.0",
+    "react-intl": "^5.8.1",
     "react-loadable": "^5.5.0",
-    "react-redux": "^7.1.3",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
+    "react-redux": "^7.2.1",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-router-redux": "^4.0.8",
-    "react-scripts": "^3.3.0",
+    "react-scripts": "^3.4.3",
     "react-text-mask": "^5.4.3",
     "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
@@ -72,23 +72,23 @@
     "store": "^2.0.12"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-decorators": "^7.7.4",
-    "@babel/preset-flow": "^7.7.4",
+    "@babel/plugin-proposal-decorators": "^7.10.5",
+    "@babel/preset-flow": "^7.10.4",
     "babel-cli": "^6.26.0",
     "babel-plugin-import": "^1.13.0",
     "babel-preset-flow": "^6.23.0",
-    "customize-cra": "^0.4.1",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-prettier": "^4.2.0",
-    "husky": "^3.1.0",
-    "less": "^3.10.3",
-    "less-loader": "^5.0.0",
-    "lint-staged": "^9.5.0",
-    "node-sass": "^4.13.0",
-    "react-app-rewired": "^2.1.5",
-    "stylelint": "^10.1.0",
-    "stylelint-config-prettier": "^5.3.0",
-    "stylelint-config-standard": "^18.3.0"
+    "customize-cra": "^1.0.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-prettier": "^6.11.0",
+    "husky": "^4.3.0",
+    "less": "^3.12.2",
+    "less-loader": "^7.0.1",
+    "lint-staged": "^10.3.0",
+    "node-sass": "^4.14.1",
+    "react-app-rewired": "^2.1.6",
+    "stylelint": "^13.7.1",
+    "stylelint-config-prettier": "^8.0.2",
+    "stylelint-config-standard": "^20.0.0"
   },
   "husky": {
     "hooks": {
